+++ packages/config/src/__test__/parse.test.ts
@@ -31,4 +31,25 @@ describe("parse", () => {
   it("returns the fallback value as a string", () => {
     expect(parse(undefined, "abc")).toBe("abc");
   });
+
+  it("returns undefined", () => {
+    // eslint-disable-next-line unicorn/no-useless-undefined
+    expect(parse(undefined, undefined)).toBe(undefined);
+  });
+
+  it("throws SyntaxError Exception due to json parse on boolean", () => {
+    try {
+      parse("Dzango", false);
+    } catch (error) {
+      expect(error).toBeInstanceOf(SyntaxError);
+    }
+  });
+
+  it("returns SyntaxError Exception due to json parse on number", () => {
+    try {
+      parse("Dzango", 14);
+    } catch (error) {
+      expect(error).toBeInstanceOf(SyntaxError);
+    }
+  });
 });