+++ packages/slonik/src/index.ts
@@ -1,10 +1,7 @@
+import { sql } from "slonik";
+
 import type { SlonikConfig } from "./types";
-import type { DatabasePool } from "slonik";
-import type {
-  ConnectionRoutine,
-  QueryFunction,
-  SqlTaggedTemplate,
-} from "slonik/dist/src/types";
+import type { ConnectionRoutine, DatabasePool, QueryFunction } from "slonik";
 
 declare module "fastify" {
   interface FastifyInstance {
@@ -13,7 +10,7 @@ declare module "fastify" {
       pool: DatabasePool;
       query: QueryFunction;
     };
-    sql: SqlTaggedTemplate<Record<never, never>>;
+    sql: typeof sql;
   }
 
   interface FastifyRequest {
@@ -22,7 +19,7 @@ declare module "fastify" {
       pool: DatabasePool;
       query: QueryFunction;
     };
-    sql: SqlTaggedTemplate<Record<never, never>>;
+    sql: typeof sql;
   }
 }
 
@@ -50,6 +47,7 @@ export { default as DefaultSqlFactory } from "./sqlFactory";
 export type {
   Database,
   FilterInput,
+  PaginatedList,
   Service,
   SlonikConfig,
   SortInput,