+++ packages/user/src/model/user-profiles/service.ts
@@ -20,9 +20,12 @@ class UserProfileService<
     Service<UserProfile, UserProfileCreateInput, UserProfileUpdateInput>
 {
   /* eslint-enabled */
-  static readonly TABLE = "users";
   static readonly LIMIT_DEFAULT = 20;
   static readonly LIMIT_MAX = 50;
+
+  get table() {
+    return this.config.user?.table?.name || "users";
+  }
 
   get factory() {
     if (!this.table) {