+++ packages/user/src/supertokens/recipes/config/third-party-email-password/emailPasswordSignUpPost.ts
@@ -6,10 +6,9 @@ import type { APIInterface } from "supertokens-node/recipe/thirdpartyemailpasswo
 
 const emailPasswordSignUpPOST = (
   originalImplementation: APIInterface,
+  /* eslint-disable-next-line @typescript-eslint/no-unused-vars */
   fastify: FastifyInstance
-): typeof originalImplementation.emailPasswordSignUpPOST => {
-  const { log } = fastify;
-
+): APIInterface["emailPasswordSignUpPOST"] => {
   return async (input) => {
     if (originalImplementation.emailPasswordSignUpPOST === undefined) {
       throw new Error("Should never come here");
@@ -19,15 +18,6 @@ const emailPasswordSignUpPOST = (
       await originalImplementation.emailPasswordSignUpPOST(input);
 
     if (originalResponse.status === "OK") {
-      const rolesResponse = await UserRoles.addRoleToUser(
-        originalResponse.user.id,
-        "USER"
-      );
-
-      if (rolesResponse.status !== "OK") {
-        log.error(rolesResponse.status);
-      }
-
       const { roles } = await UserRoles.getRolesForUser(
         originalResponse.user.id
       );